Bake multi-query into SRequest to get support for ensureUnique
