+++ src/main/java/dk/kb/netarchivesuite/solrwayback/solr/SolrGenericStreaming.java
@@ -32,13 +32,11 @@ import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.Set;
 import java.util.Spliterators;
-import java.util.regex.Pattern;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
@@ -104,35 +102,21 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
    * Default page size (rows) for the cursormark paging.
    */
   public static final int DEFAULT_PAGESIZE = 1000;
-
   /**
-   * Default sort used when exporting. Ends with tie breaking on id.
+   * Used for batching multiple queries. Newer Solrs (at least 9+) share a default upper limit of 1024 boolean clauses
+   * recursively in the user issued query tree.
    */
-  public static final String DEFAULT_SORT = "score desc, id asc";
+  public static final int DEFAULT_QUERY_BATCHSIZE = 200;
 
-  /**
-   * Default maximum number of elements when the request requires unique results.
-   * If this limit is exceeded during processing, an exception is thrown.
-   * The uniquifier uses a HashSet (which is a Map underneath the hood) for tracking
-   * unique values. Each entry takes up about ~150 bytes plus the value itself, so
-   * something like 250 bytes/entry as a rule of thumb. The default MAX_UNIQUE is thus
-   * about 1.25GB of maximum heap.
-   */
-  public static final int DEFAULT_MAX_UNIQUE = 5_000_000;
   /**
    * Magic value for signalling that there are no more pages.
    */
   public static final String STOP_PAGING = "___STOP_PAGING___";
 
-  /**
-   * Solr ISO timestamp parsing. Supports optional milliseconds.
-   * Sample inputs: {@code 2022-09-26T12:05:00Z}, {@code 2022-09-26T12:05:00.123Z}.
-   */
-  private static final Pattern ISO_TIME = Pattern.compile(
-          "[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]T[012][0-9]:[0-5][0-9]:[0-5][0-9][.]?[0-9]?[0-9]?[0-9]?Z");
-
   private final SRequest request;
-  private final SolrQuery solrQuery; // Constructed from request
+  private final SolrQuery solrQuery; // Constructed from request. Used for single-query requests
+  private SolrQuery solrBatchQuery = null; // Used for multi-query requests
+  private final Iterator<String> queries; // Only defined if the request is multi-query
 
   private final List<String> initialFields;  // Caller provided fields
   private final List<String> adjustedFields; // Fields after adjusting for unique etc.
@@ -149,7 +133,7 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
   /**
    * The default SolrClient is simple and non-caching as streaming exports typically makes unique requests.
    */
-  private static SolrClient defaultSolrClient = new HttpSolrClient.Builder(PropertiesLoader.SOLR_SERVER).build();
+  static SolrClient defaultSolrClient = new HttpSolrClient.Builder(PropertiesLoader.SOLR_SERVER).build();
 
 
   /**
@@ -180,37 +164,6 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
   }
 
   /**
-   * Issues multiple queries as batch requests. The individual queries will be {@code OR}ed together:
-   *
-   * If {@code queries = Arrays.asList("url:http://example.com/foo", "url:http://example.com/bar").stream()},
-   * the batch request will be {@code "url:http://example.com/foo OR url:http://example.com/bar"}.
-   *
-   * Each batch request is independent from previous requests so de-duplication and uniqueness is not guaranteed to
-   * work well.
-   *
-   * Sample call: Get all SolrDocuments matching the queries {@code foo} and {@code bar}:
-   * <pre>
-   * Stream<SolrDocument> solrDocs =
-   *   multiQuery(SRequest.builder().fields("id"),
-   *              Stream.of("foo", "var"),
-   *              10).
-   *       flatMap(SolrGenericStreaming::stream);
-   * </pre>
-   * @param baseRequest the SRequest used as a base for each batch request: {@link SRequest#solrQuery} will be
-   *                    overridden with the batch queries constructed from queries.
-   * @param queries     the individual queries. These will be concatenated in batches with {@code " OR "} as separator.
-   * @param batchSize   the number of queries to use for each batch. Setting this above 1000 is not likely to work
-   *                    due to Solr's default of {@code maxBooleanClauses = 1024}.
-   * @return a stream of instances of {@link SolrGenericStreaming}.
-   */
-  public static Stream<SolrGenericStreaming> multiQuery(SRequest baseRequest, Stream<String> queries, int batchSize) {
-    return CollectionUtils.splitToLists(queries, batchSize).
-            map(batch -> "(" + String.join(") OR (", batch) + ")"). // 1 big OR query
-            map(query -> baseRequest.deepCopy().query(query)).
-            map(SolrGenericStreaming::create);
-  }
-
-  /**
    * Generic stream where all parts except {@link SRequest#query(String)} and {@link SRequest#fields(String...)}
    * are optional.
    *
@@ -225,15 +178,20 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
     this.adjustedFields = Arrays.asList(solrQuery.getFields().split(","));
 
     this.uniqueTracker = request.ensureUnique ? new HashSet<>() : null;
+    queries = request.queries == null ? null:
+            CollectionUtils.splitToLists(request.queries, request.queryBatchSize).
+                    map(batch -> "(" + String.join(") OR (", batch) + ")").
+                    iterator(); // 1 big OR query
   }
 
   /**
    * Adjust the given solrQuery so that it will work for streaming processing.
+   * This is called automatically when using {@link SRequest}s.
    *
    * If {@code fl} is not already set in solrQuery it will be set to {@code source_file_path,source_file_offset}.
    * If {@code cursorMark} is not already set in solrQuery it will be set to {@code *}.
    * If {@code rows} is not already set in solrQuery it will be set to {@link #DEFAULT_PAGESIZE}.
-   * If {@code sort} is not already set in solrQuery it will be set to {@link #DEFAULT_SORT}.
+   * If {@code sort} is not already set in solrQuery it will be set to {@link SRequest#DEFAULT_SORT}.
    * If {@code sort} does not end with {@code id asc} or {@code id desc}, {@code id asc} will be appended.
    * If expandResources is true and {@code fl} in solrQuery does not contain {@code content_type_norm},
    * {code source_file_path} and {@code source_file_offset} they will be added.
@@ -286,7 +244,7 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
     solrQuery.set(CommonParams.FL, String.join(",", fl));
 
     // Adjust sort to ensure presence of tie breaker and deduplication field (if enabled)
-    String sort = solrQuery.get(CommonParams.SORT, DEFAULT_SORT);
+    String sort = solrQuery.get(CommonParams.SORT, SRequest.DEFAULT_SORT);
     if (!(sort.endsWith("id asc") || sort.endsWith("id desc"))) {
       sort = sort + ", id asc"; // A tie breaker is needed when using cursormark
     }
@@ -393,23 +351,41 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
    * @throws IOException if general communication with Solr failed.
    */
   public SolrDocumentList nextDocuments() throws SolrServerException, IOException {
-    String cursorMark = solrQuery.get(CursorMarkParams.CURSOR_MARK_PARAM, CursorMarkParams.CURSOR_MARK_START);
     while (!hasFinished && delivered < request.maxResults) {
-
       // Return batch if undelivered contains any documents
       if (undelivered != null && undelivered.size() > 0) {
         return nextPageUndelivered();
       }
 
+      SolrQuery solrRealQuery = solrQuery;
+      if (request.isMultiQuery() && solrBatchQuery == null) { // Multi-query
+        if (!queries.hasNext()) {
+          hasFinished = true;
+          return null;
+        }
+        solrBatchQuery = SolrUtils.deepCopy(solrQuery).setQuery(queries.next());
+        solrBatchQuery.set(CursorMarkParams.CURSOR_MARK_PARAM,
+                           solrQuery.get(CursorMarkParams.CURSOR_MARK_PARAM, CursorMarkParams.CURSOR_MARK_START));
+        solrRealQuery = solrBatchQuery;
+      }
+
+
+      String cursorMark = solrRealQuery.get(CursorMarkParams.CURSOR_MARK_PARAM, CursorMarkParams.CURSOR_MARK_START);
+
       // Only page to next page if there are more pages
       if (STOP_PAGING.equals(cursorMark)) {
+        if (request.isMultiQuery()) {
+          // Multi-query, so skip to next query
+          solrBatchQuery = null;
+          continue;
+        }
         hasFinished = true;
         return null;
       }
 
       // No more documents buffered, attempt to require new documents
-      solrQuery.set(CursorMarkParams.CURSOR_MARK_PARAM, cursorMark);
-      QueryResponse rsp = request.solrClient.query(solrQuery, METHOD.POST);
+      solrRealQuery.set(CursorMarkParams.CURSOR_MARK_PARAM, cursorMark);
+      QueryResponse rsp = request.solrClient.query(solrRealQuery, METHOD.POST);
       undelivered = rsp.getResults();
 
       if (request.deduplicateField != null) {
@@ -436,7 +412,7 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
       } else {
         cursorMark = newCursormark;
       }
-      solrQuery.set(CursorMarkParams.CURSOR_MARK_PARAM, cursorMark);
+      solrRealQuery.set(CursorMarkParams.CURSOR_MARK_PARAM, cursorMark);
 
       // Loop as deduplication & unique might mean that the current batch is empty
     }
@@ -604,308 +580,6 @@ public class SolrGenericStreaming implements Iterable<SolrDocument> {
               return new SolrDocument(extraMap);
             }).
             flatMap(SolrGenericStreaming::flatten);
-  }
-
-  /**
-   * Encapsulation of the request to SolrGenericStreaming. Care has been taken to ensure sane defaults; the only
-   * required attributes are {@link #query(String)} and {@link #fields(String...)}.
-   * 
-   * Use as a builder: {@code SRequest.builder().query("*:*").fields("url", "url_norm")}
-   * 
-   * Note: If {@link SRequest#solrQuery(SolrQuery)} is specified it will be used as a base for the full request.
-   */
-  public static class SRequest {
-    public SolrClient solrClient = defaultSolrClient;
-    public SolrQuery solrQuery = new SolrQuery();
-    public boolean expandResources = false;
-    public boolean ensureUnique = false;
-    public Integer maxUnique = DEFAULT_MAX_UNIQUE;
-    private String idealTime; // If defined, a sort will be created as String.format(Locale.ROOT, "%s asc, abs(sub(ms(%s), crawl_date)) asc", deduplicateField, idealTime);
-    public String deduplicateField = null;
-    List<String> fields;
-    public long maxResults = Long.MAX_VALUE;
-    public String sort = DEFAULT_SORT;
-    public String query = null;
-    public List<String> filterQueries;
-    public int pageSize = DEFAULT_PAGESIZE;
-
-    /**
-     * @return a fresh instance of SRequest intended for further adjustment.
-     */
-    public static SRequest builder() {
-      return new SRequest();
-    }
-
-    /**
-     * Creates a request with query and fields.
-     * As the returned request is initialized with query and fields, it can be used without further adjustments.
-     * @return an instance of SRequest, initialized with the provided query and fields.
-     */
-    public static SRequest create(String query, String... fields) {
-      return new SRequest().query(query).fields(fields);
-    }
-
-    /**
-     * Creates a request with query and fields.
-     * As the returned request is initialized with query and fields, it can be used without further adjustments.
-     * @return an instance of SRequest, initialized with the provided query and fields.
-     */
-    public static SRequest create(String query, List<String> fields) {
-      return new SRequest().query(query).fields(fields);
-    }
-
-    /**
-     * @param solrClient       used for issuing Solr requests. If not specified, {@link #defaultSolrClient} will be used.
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest solrClient(SolrClient solrClient) {
-      if (solrClient == null) {
-        log.debug("solrClient(null) called. Leaving solrClient unchanged");
-        return this;
-      }
-      this.solrClient = solrClient;
-      return this;
-    }
-
-    /**
-     * The parameters in the solrQuery has the lowest priority: All calls to modifier methods will override matching
-     * values in the solrQuery.
-     * @param solrQuery        the base for the request. If not provided it will be constructed from scratch.
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest solrQuery(SolrQuery solrQuery) {
-      this.solrQuery = solrQuery;
-      return this;
-    }
-
-    /**
-     * @param expandResources  if true, embedded resources for HTML pages are extracted and added to the delivered
-     *                         lists of Solr Documents. Default is false.
-     *                         Note: Indirect references (through JavaScript & CSS) are not followed.
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest expandResources(boolean expandResources) {
-      this.expandResources = expandResources;
-      return this;
-    }
-
-    /**
-     * @param ensureUnique     if true, unique documents are guaranteed. This is only sane if expandResources is true.
-     *                         Default is false.
-     *                         Note that a HashSet is created to keep track of encountered documents and will impose
-     *                         a memory overhead linear to the number of results.
-     * @return the SRequest adjusted with the provided value.
-     * @see #maxUnique(Integer)
-     */
-    public SRequest ensureUnique(boolean ensureUnique) {
-      this.ensureUnique = ensureUnique;
-      return this;
-    }
-
-    /**
-     * @param maxUnique        the maximum number of uniques to track when ensureUnique is true.
-     *                         If the number of uniques exceeds this limit, an exception will be thrown.
-     *                         Default is {@link #DEFAULT_MAX_UNIQUE}.
-     * @return the SRequest adjusted with the provided value.
-     * @see #ensureUnique(boolean)
-     */
-    public SRequest maxUnique(Integer maxUnique) {
-      this.maxUnique = maxUnique;
-      return this;
-    }
-
-    /**
-     * @param deduplicateField The field to use for de-duplication. This is typically {@code url}.
-     *                         Default is null (no deduplication).
-     *                         Note: deduplicateField does not affect expandResources. Set ensureUnique to true if
-     *                         if expandResources is true and uniqueness must also be guaranteed for resources.
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest deduplicateField(String deduplicateField) {
-      this.deduplicateField = deduplicateField;
-      return this;
-    }
-
-    /**
-     * Deduplication combined with time proximity sorting. This is a shorthand for
-     * {@code request.deduplicatefield(deduplicateField).sort("abs(sub(ms(idealTime), crawl_date)) asc");}
-     *
-     * Use case: Extract unique URLs for matches that are closest to a given point in time:
-     *           {@code timeProximityDeduplication("2014-01-03T11:56:58Z", "url_norm"}.
-     *
-     * Note: This overrides any existing {@link #sort(String)}.
-     * @param idealTime        The time that the resources should be closest to, stated as a Solr timestamp
-     *                         {@code YYYY-MM-DDTHH:mm:SSZ}.
-     *                         Also supports {@code oldest} and {@code newest} as values.
-     * @param deduplicateField The field to use for de-duplication. This is typically {@code url}.
-     * @return the SRequest adjusted with the provided values.
-     */
-    public SRequest timeProximityDeduplication(String idealTime, String deduplicateField) {
-      String origo = idealTime;
-      if ("newest".equals(idealTime)) {
-        origo = "9999-12-31T23:59:59Z";
-      } else if ("oldest".equals(idealTime)) {
-        origo = "0001-01-01T00:00:01Z";
-      } else if (!ISO_TIME.matcher(idealTime).matches()) {
-        throw new IllegalArgumentException(
-                "The idealTime '" + idealTime + "' does not match 'oldest', 'newest', 'YYYY-MM-DDTHH:mm:SSZ' or " +
-                "'YYYY-MM-DDTHH:mm:SS.sssZ");
-      }
-      this.idealTime = origo;
-
-      if (deduplicateField == null) {
-        throw new NullPointerException("deduplicateField == null which is not allowed for timeProximityDeduplication");
-      }
-      this.deduplicateField = deduplicateField;
-      return this;
-    }
-
-    /**
-     * @param fields           fields to export (fl). deduplicateField will be added to this is not already present.
-     *                         This parameter has no default and must be defined.
-     * @return the SRequest adjusted with the provided value.
-     * @see #fields(String...)
-     */
-    public SRequest fields(List<String> fields) {
-      this.fields = fields;
-      return this;
-    }
-
-    /**
-     * @param fields           fields to export (fl). deduplicateField will be added to this is not already present.
-     *                         This parameter has no default and must be defined.
-     *                         If only 1 field is specified and fields contains at least 1 comma, it will treated
-     *                         as a comma separated list of fields: {@code "foo,bar,zoo" == "foo", "bar", "zoo"}.
-     * @return the SRequest adjusted with the provided value.
-     * @see #fields(List)
-     */
-    public SRequest fields(String... fields) {
-      if (fields.length == 1 && fields[0].contains(",")) {
-        this.fields = Arrays.asList(fields[0].split(", *"));
-      } else {
-        this.fields = Arrays.asList(fields);
-      }
-      return this;
-    }
-
-    /**
-     * @param maxResults       the maximum number of results to return. This includes expanded resources.
-     *                         Default is {@link Long#MAX_VALUE} (effectively no limit).
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest maxResults(long maxResults) {
-      this.maxResults = maxResults;
-      return this;
-    }
-
-    /**
-     * Note: {@link #timeProximityDeduplication(String, String)} takes precedence over sort.
-     * @param sort             standard Solr sort. Depending on deduplicateField and tie breaker it might be adjusted
-     *                         by {@link #adjustSolrQuery(SolrQuery, boolean, boolean, String)}.
-     *                         Default is {@link #DEFAULT_SORT}.
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest sort(String sort) {
-      this.sort = sort;
-      return this;
-    }
-
-    /**
-     * @param query            standard Solr query.
-     *                         This parameter has no default and must be defined either directly or
-     *                         through {@link #solrQuery(SolrQuery)}.
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest query(String query) {
-      this.query = query;
-      return this;
-    }
-
-    /**
-     * @param filterQueries    optional Solr filter queries. For performance, 0 or 1 filter query is recommended.
-     *                         If multiple filters are to be used, consider collapsing them into one:
-     *                         {@code ["foo", "bar"]} → {@code ["(foo) AND (bar)"]}.
-     * @return the SRequest adjusted with the provided value.
-     * @see #filterQueries(String...)
-     */
-    public SRequest filterQueries(List<String> filterQueries) {
-      this.filterQueries = filterQueries;
-      return this;
-    }
-    /**
-     * @param filterQueries    optional Solr filter queries. For performance, 0 or 1 filter query is recommended.
-     *                         If multiple filters are to be used, consider collapsing them into one:
-     *                         {@code ["foo", "bar"]} → {@code ["(foo) AND (bar)"]}.
-     * @return the SRequest adjusted with the provided value.
-     * @see #filterQueries(List)
-     */
-    public SRequest filterQueries(String... filterQueries) {
-      this.filterQueries = Arrays.asList(filterQueries);
-      return this;
-    }
-
-    /**
-     * @param pageSize         paging size. Typically 500-100,000 depending on fields.
-     *                         Default is {@link #DEFAULT_PAGESIZE}.
-     * @return the SRequest adjusted with the provided value.
-     */
-    public SRequest pageSize(int pageSize) {
-      this.pageSize = pageSize;
-      return this;
-    }
-
-    /**
-     * Copies {@link #solrQuery} and adjusts it with defined attributes from the SRequest, extending with needed
-     * SolrRequest-attributes.
-     * @return a SolrQuery ready for processing.
-     */
-    public SolrQuery getMergedSolrQuery() {
-      SolrQuery solrQuery = SolrUtils.deepCopy(this.solrQuery);
-      if (query != null) {
-        solrQuery.setQuery(query);
-      }
-      if (filterQueries != null) {
-        solrQuery.setFilterQueries(filterQueries.toArray(new String[0]));
-      }
-
-      if (idealTime != null) {
-        sort = String.format(Locale.ROOT, "%s asc, abs(sub(ms(%s), crawl_date)) asc", deduplicateField, idealTime);
-      }
-      solrQuery.set(CommonParams.SORT, sort);
-      if (fields != null) {
-        solrQuery.set(CommonParams.FL, String.join(",", fields));
-      }
-      solrQuery.set(CommonParams.ROWS, (int)Math.min(maxResults, pageSize));
-      return solrQuery;
-    }
-
-    /**
-     * @return a copy of this SRequest, as independent as possible: {@link #solrQuery} and Lists are deep-copied.
-     */
-    public SRequest deepCopy() {
-      SRequest copy = new SRequest().
-              solrClient(solrClient).
-              solrQuery(solrQuery == null ? null : SolrUtils.deepCopy(solrQuery)).
-              expandResources(expandResources).
-              ensureUnique(ensureUnique).
-              maxUnique(maxUnique).
-              deduplicateField(deduplicateField).
-              fields(copy(fields)).
-              maxResults(maxResults).
-              sort(sort).
-              query(query).
-              filterQueries(copy(filterQueries)).
-              pageSize(pageSize);
-      copy.idealTime = idealTime;
-      return copy;
-    }
-
-    private List<String> copy(List<String> fields) {
-      if (fields == null) {
-        return null;
-      }
-      return new ArrayList<>(fields);
-    }
   }
 
 }
